Change codesign_app.sh to allow ad-hoc codesigning
Adding a second param to `codesign_app.sh` to provide signing identity.
An identity can be created with a self-signed certificate in Keychain Access.app -> Certificate Assistant -> Create a Certificate.
For non-AppStore distribution, an Apple-issued "Developer ID" certificate identity can be provided. Finally, if the param is omitted, the ad-hoc signing is valid only for current machine where the signing happened.
ad-hoc codesigned app does not use hardened runtime or entitlements, and does not use timestamp (which is needed for distribution)