Heuristics to approximate the best context size for available RAM
It would be better to parse gguf metadata and make a better estimate but in the meantime this should be better than default 4096 tokens. Wish llama-server could choose the optimal size automatically. It looks like it does some validation of the number because for some models it lowers it down if our estimate comes with too high number.